+++ omnisend/class-omnisend-core-bootstrap.php
@@ -19,6 +19,7 @@
  */
 
 use Omnisend\Internal\Options;
+use Omnisend\Internal\Connection;
 
 defined( 'ABSPATH' ) || die( 'no direct access' );
 
@@ -154,7 +155,50 @@ class Omnisend_Core_Bootstrap {
 	}
 
 	public static function load_react(): void {
-		
+		if ( Connection::show_connected_store_view() ) {
+			add_action(
+				'admin_enqueue_scripts',
+				function ( $suffix ) {
+					$asset_file_page = plugin_dir_path( __FILE__ ) . 'build/connected.asset.php';
+					if ( file_exists( $asset_file_page ) && 'toplevel_page_omnisend' === $suffix ) {
+						$assets = require_once $asset_file_page;
+						wp_enqueue_script(
+							'connected-script',
+							plugin_dir_url( __FILE__ ) . 'build/connected.js',
+							$assets['dependencies'],
+							$assets['version'],
+							true
+						);
+						foreach ( $assets['dependencies'] as $style ) {
+							wp_enqueue_style( $style );
+						}
+					}
+				}
+			);
+		}
+
+		if ( Connection::show_connection_view() ) {
+			add_action(
+				'admin_enqueue_scripts',
+				function ( $suffix ) {
+					$asset_file_page = plugin_dir_path( __FILE__ ) . 'build/connection.asset.php';
+					if ( file_exists( $asset_file_page ) ) {
+						$assets = require_once $asset_file_page;
+						wp_enqueue_script(
+							'connection-script',
+							plugin_dir_url( __FILE__ ) . 'build/connection.js',
+							$assets['dependencies'],
+							$assets['version'],
+							true
+						);
+						foreach ( $assets['dependencies'] as $style ) {
+							wp_enqueue_style( $style );
+						}
+					}
+				}
+			);
+		}
+
 		add_action(
 			'admin_enqueue_scripts',
 			function ( $suffix ) {
@@ -173,7 +217,7 @@ class Omnisend_Core_Bootstrap {
 					}
 				}
 			}
-		);	
+		);
 	}
 
 	public static function is_omnisend_woocommerce_plugin_active(): bool {
